requirements.txt
asgiref==3.8.1
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
